Move low level propagation tests
